package's refactoring and add logger
