+++ src/main/java/com/jsonproject/finder/Main.java
@@ -1,10 +1,16 @@
 package com.jsonproject.finder;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
 public class Main {
 
 
     public static void main(String[] args) {
-        //write hello world
+
+    Logger logger = LogManager.getRootLogger();
+
+    logger.info("123 аврапр");
 
 
     }
