+++ src/main/java/com/jsonproject/finder/entity/TaxiDriver.java
@@ -1,4 +1,4 @@
-package com.jsonproject.finder.entities;
+package com.jsonproject.finder.entity;
 
 import lombok.*;
