+++ src/main/resources/log4j2.properties
@@ -1,0 +1,20 @@
+status = info
+name = PropertiesConfig
+
+#console appender
+appender.console.type = Console
+appender.console.name = ConsoleAppender
+appender.console.layout.type = PatternLayout
+appender.console.layout.pattern = %d{HH:mm:ss.SSS} [%t] %highlight{%-5level}{FATAL=red blink, ERROR=red, WARN=yellow, INFO=green, DEBUG=white, TRACE=white} %logger{36} - %msg%n
+appender.console.layout.charset = UTF-8
+
+
+#root logger settings
+# all, trace, debug, info, warn, error, fatal, off
+rootLogger.level = info
+rootLogger.appenderRefs = console
+rootLogger.appenderRef.console.ref = ConsoleAppender
+
+
+
+
